CiviLens_backend/requirements.txt
Django>=4.2
djongo
pymongo
python-dotenv
PyJWT
django-cors-headers
